Changes lick reporting to diff of pre-post baseline
In order for an animal to get a reward the difference between the number of
licks before the stimulus. The numbers are averaged and there is floating
point arithmetic thrown in as well so this might be slow.

Also:
Adjusted the counting in `count_responses`. It now more intuitivley accpets
a single argument which is the number of milliseconds to count licks over.